+++ social_media.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+"""Unit providing simple interfaces for social media accounts I use.
+
+Imports another unit, social_media_auth.py, that contains the necessary
+authentication constants. social_media_auth.py is not, of course, contained
+in the GitHub repo that contains this unit, because, well, those are my private
+authentication constants."""
+
+from tumblpy import Tumblpy
+
+if __name__ == "__main__":
+    pass